Added dev chains

+++ src/Node/GethNode.ts
@@ -4,6 +4,7 @@ import Node from './Node';
 import Shell from '../Shell';
 import Directory from '../Directory';
 
+const DEV_CHAIN_DOCKER = 'deepeshkn/test_docker_image';
 /**
  * Represents a geth node that runs in a docker container.
  */
@@ -18,8 +19,13 @@ export default class GethNode extends Node {
     this.initializeDirectories();
     super.ensureNetworkExists();
 
+    let args = [];
+    if (this.chain === 'dev' || this.originChain === 'dev') {
+      args = this.devGethArgs(this.chain);
+    } else {
+      args = this.defaultDockerGethArgs;
+    }
     this.logInfo('starting geth node');
-    const args = this.defaultDockerGethArgs;
     Shell.executeDockerCommand(args);
   }
 
@@ -65,7 +71,7 @@ export default class GethNode extends Node {
     }
   }
 
-  private get defaultDockerGethArgs(): string[] {
+  private getDefaultDockerArgs(): string[] {
     let args = [
       'run',
     ];
@@ -81,15 +87,21 @@ export default class GethNode extends Node {
       '--publish', `${this.port}:${this.port}`,
       '--publish', `${this.rpcPort}:8545`,
       '--publish', `${this.websocketPort}:8546`,
-      '--volume', `${this.chainDir}:/chain_data`,
     ]);
 
+    return args;
+  }
+
+  private get defaultDockerGethArgs(): string[] {
+    let args = this.getDefaultDockerArgs();
+    args = args.concat([
+      '--volume', `${this.chainDir}:/chain_data`,
+    ]);
     if (this.password !== '') {
       args = args.concat([
         '--volume', `${this.password}:/password.txt`,
       ]);
     }
-
     args = args.concat([
       'ethereum/client-go:v1.8.23',
       '--networkid', this.chain,
@@ -128,6 +140,26 @@ export default class GethNode extends Node {
     return args;
   }
 
+  private devGethArgs(chain): string[] {
+    let args = this.getDefaultDockerArgs();
+    let volume = '';
+
+    if (chain === 'dev' || chain === 'origin') {
+      volume = `${this.chainDir}:/origin_volume`;
+    } else {
+      volume = `${this.chainDir}:/auxiliary_volume`;
+    }
+    args = args.concat([
+      '--volume', volume,
+    ]);
+    args = args.concat([
+      DEV_CHAIN_DOCKER,
+      chain === 'dev' || chain === 'origin' ? 'origin' : 'auxiliary',
+    ]);
+
+    return args;
+  }
+
   /**
    * Copies the initialized geth repository to the data directory if it does not exist.
    */
@@ -136,17 +168,20 @@ export default class GethNode extends Node {
 
     if (!fs.existsSync(this.chainDir)) {
       this.logInfo(`${this.chainDir} does not exist; initializing`);
-      fs.mkdirSync(this.chainDir);
-      fs.copySync(
-        path.join(
-          Directory.projectRoot,
-          'chains',
-          this.originChain,
-          this.chain,
-          'geth',
-        ),
-        path.join(this.chainDir, 'geth'),
+      fs.mkdirpSync(this.chainDir);
+      const sourcePath = path.join(
+        Directory.projectRoot,
+        'chains',
+        this.originChain,
+        this.chain,
+        'geth',
       );
+      if (fs.existsSync(sourcePath)) {
+        fs.copySync(
+          sourcePath,
+          path.join(this.chainDir, 'geth'),
+        );
+      }
     }
   }
 }
